requirements.txt
asgiref==3.8.1
boto3==1.36.7
botocore==1.36.7
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.5
django-storages==1.14.4
fonttools==4.55.7
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
psycopg==3.2.4
psycopg2==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
XlsxWriter==3.2.1
django-environ==0.11.2